+++ src/main/kotlin/me/partlysanestudios/partlysaneskies/PartlySaneSkies.kt
@@ -79,8 +79,8 @@ import me.partlysanestudios.partlysaneskies.features.gui.RefreshKeybinds
 import me.partlysanestudios.partlysaneskies.features.gui.hud.CooldownHud
 import me.partlysanestudios.partlysaneskies.features.gui.hud.LocationBannerDisplay
 import me.partlysanestudios.partlysaneskies.features.gui.hud.rngdropbanner.DropBannerDisplay
-import me.partlysanestudios.partlysaneskies.features.items.rndropbanner.DropWebhook
-import me.partlysanestudios.partlysaneskies.features.items.rndropbanner.RareDropGUIManager
+import me.partlysanestudios.partlysaneskies.features.items.rngdrop.DropWebhook
+import me.partlysanestudios.partlysaneskies.features.items.rngdrop.RareDropGUIManager
 import me.partlysanestudios.partlysaneskies.features.gui.mainmenu.PSSMainMenu
 import me.partlysanestudios.partlysaneskies.features.information.WikiArticleOpener
 import me.partlysanestudios.partlysaneskies.features.mining.PickaxeWarning
