--- src/main/kotlin/me/partlysanestudios/partlysaneskies/features/items/rndropbanner/RareDropPreset.kt
@@ -1,3 +1,0 @@
-package me.partlysanestudios.partlysaneskies.features.items.rndropbanner
-
-data class RareDropPreset(val name: String, val items: List<String>)
