+++ src/main/kotlin/me/partlysanestudios/partlysaneskies/features/items/rngdrop/RareDropPreset.kt
@@ -1,0 +1,3 @@
+package me.partlysanestudios.partlysaneskies.features.items.rngdrop
+
+data class RareDropPreset(val name: String, val items: List<String>)
